refactor: export run action handler logic
